# src/main/java/demo/toll/domain/TollTxnSummary.java
package demo.toll.domain;

import java.io.Serializable;
import java.util.Date;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Data
@AllArgsConstructor
@NoArgsConstructor
public class TollTxnSummary implements Serializable {

	private static final long serialVersionUID = 1L;

	@Id
	@GeneratedValue(strategy = GenerationType.AUTO)
	private Long txnSummaryId;
	
	private Long tollPlazaId;
	private Long laneId;
	
	private String vehicleSeqNumber;
	
	private Date txnStartTime;
	private Date txnEndTime;
	
	private Long txnCount;
	private Double tollAmount;
	
}